backend/pyproject.toml
[project]
name = "lesjardinsdudore"
version = "0.0.0"
description = "Les Jardins du Dore"
authors = [
    {name = "Marc Tardif", email = "marc@taram.ca"},
]
readme = "README.rst"
repository = "https://github.com/cr3/lesjardinsdudore"
include = [
  "lesjardinsdudore/index.html",
  "permaculture/locales/*/LC_MESSAGES/*.mo",
]
requires-python = ">=3.13,<4.0"
dependencies = [
  "appdirs>=1.4.4,<2.0.0",
  "attrs>=25.3.0,<26.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "more-itertools>=10.5.0,<11.0.0",
  "permaculture>=0.0.13,<1.0.0",
  "uvicorn>=0.34.0,<1.0.0",
]

[project.optional-dependencies]

test = [
  "coverage>=7.2.3,<8.0.0",
  "httpx>=0.28.1,<1.0.0",
  "pyhamcrest>=2.1.0,<3.0.0",
  "pytest>=9.0.0,<10.0.0",
  "pytest-unique>=0.1.7,<1.0.0",
  "pytest-xdocker>=0.2.7,<1.0.0",
  "responses>=0.26.0,<1.0.0",
]

check = [
  "ruff>=0.15.0,<1.0.0",
]

docs = [
  "sphinx>=6.1.3,<7.0.0",
  "sphinxcontrib-log-cabinet>=1.0.1,<2.0.0",
  "sphinx-rtd-theme>=1.2.0,<2.0.0",
]

[project.entry-points."pytest11"]
lesjardinsdudore = "lesjardinsdudore.testing.fixtures"
lesjardinsdudore-services = "lesjardinsdudore.testing.services"

[project.entry-points."permaculture.ingestors"]
json = "lesjardinsdudore.testing.ingestor:JSONIngestor"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 79
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Create your own exception
    "TRY002",
    # Avoid specifying long messages outside the exception class
    "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

# Pytest options:
#   https://docs.pytest.org/en/6.2.x/reference.html#ini-options-ref
[tool.pytest.ini_options]
addopts = [
    "--doctest-modules",
    "--doctest-glob=*.rst",
]
testpaths = [
    "lesjardinsdudore",
    "docs",
    "tests",
]

# Coverage options:
#   https://coverage.readthedocs.io/en/latest/config.html
[tool.coverage.paths]
source = [
    "lesjardinsdudore",
    "*/*/site-packages",
]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.coverage.run]
branch = true
parallel = true
source = [
    "lesjardinsdudore",
]
